fix(pagination): address code review comments from PR #11
- toSnakeCase: handle initialisms correctly (GuardianID→guardian_id)
- parseFlexibleBool: return error for unrecognized boolean strings instead of silently defaulting to false
- page/limit validation: return 400 Bad Request for invalid/non-positive values
- FindByType: add ORDER BY for deterministic pagination
- Swagger: update @Success annotations to reflect paginated envelope, add @Param page/limit
- Fix int64→int cast in service layer (school, user, membership) to match updated shared repo interface
- Update mock repositories to return int64 total count to satisfy updated shared repo interface

+++ internal/application/service/school_service_test.go
@@ -250,7 +250,7 @@ func TestSchoolService_ListSchools(t *testing.T) {
 		{
 			name: "success - returns list",
 			setupMock: func(m *mock.MockSchoolRepository) {
-				m.ListFn = func(_ context.Context, _ sharedrepo.ListFilters) ([]*entities.School, int, error) {
+				m.ListFn = func(_ context.Context, _ sharedrepo.ListFilters) ([]*entities.School, int64, error) {
 					return []*entities.School{
 						{ID: uuid.New(), Name: "School 1", Code: "S1"},
 						{ID: uuid.New(), Name: "School 2", Code: "S2"},
@@ -263,7 +263,7 @@ func TestSchoolService_ListSchools(t *testing.T) {
 		{
 			name: "success - empty list",
 			setupMock: func(m *mock.MockSchoolRepository) {
-				m.ListFn = func(_ context.Context, _ sharedrepo.ListFilters) ([]*entities.School, int, error) {
+				m.ListFn = func(_ context.Context, _ sharedrepo.ListFilters) ([]*entities.School, int64, error) {
 					return []*entities.School{}, 0, nil
 				}
 			},
@@ -273,7 +273,7 @@ func TestSchoolService_ListSchools(t *testing.T) {
 		{
 			name: "error - database error",
 			setupMock: func(m *mock.MockSchoolRepository) {
-				m.ListFn = func(_ context.Context, _ sharedrepo.ListFilters) ([]*entities.School, int, error) {
+				m.ListFn = func(_ context.Context, _ sharedrepo.ListFilters) ([]*entities.School, int64, error) {
 					return nil, 0, fmt.Errorf("timeout")
 				}
 			},
